Fee compound

+++ contracts/test/strategies/TestLiquidationStrategy.sol
@@ -39,6 +39,6 @@ contract TestLiquidationStrategy is ILiquidationStrategy  {
         refund[0] = 13;
         refund[1] = 14;
         emit Liquidation(0, tokensHeld[0], tokensHeld[1], uint128(totalLoanLiquidity), TX_TYPE.BATCH_LIQUIDATION, tokenIds);
-        emit PoolUpdated(totalCollateral, refund[0], uint96(refund[1]), 700, 800, 900, 1000, TX_TYPE.BATCH_LIQUIDATION);
+        emit PoolUpdated(totalCollateral, refund[0], uint48(refund[1]), 700, 800, 900, 1000, TX_TYPE.BATCH_LIQUIDATION);
     }
 }
